Use conda to build python packages during GPU tests
