+++ ci/gpu/build.sh
@@ -30,6 +30,7 @@ cd $WORKSPACE
 # Parse git describe
 export GIT_DESCRIBE_TAG=`git describe --tags`
 export MINOR_VERSION=`echo $GIT_DESCRIBE_TAG | grep -o -E '([0-9]+\.[0-9]+)'`
+unset GIT_DESCRIBE_TAG
 
 # ucx-py version
 export UCX_PY_VERSION='0.26.*'
@@ -169,6 +170,7 @@ else
     gpuci_logger "Check GPU usage"
     nvidia-smi
 
+    gpuci_logger "Installing libcuml and libcuml-tests"
     gpuci_mamba_retry install -y -c ${CONDA_ARTIFACT_PATH} libcuml libcuml-tests
 
     gpuci_logger "Running libcuml test binaries"
@@ -180,20 +182,11 @@ else
         echo "Ran gtest $test_name : return code was: $?, test script exit code is now: $EXITCODE"
     done
 
-    # FIXME: Project FLASH only builds for python version 3.8 which is the one used in
-    # the CUDA 11.0 job, need to change all versions to project flash
-    if [ "$CUDA_REL" == "11.0" ];then
-        gpuci_logger "Using Project FLASH to install cuml python"
-        CONDA_FILE=`find ${CONDA_ARTIFACT_PATH} -name "cuml*.tar.bz2"`
-        CONDA_FILE=`basename "$CONDA_FILE" .tar.bz2` #get filename without extension
-        CONDA_FILE=${CONDA_FILE//-/=} #convert to conda install
-        echo "Installing $CONDA_FILE"
-        gpuci_mamba_retry install -c ${CONDA_ARTIFACT_PATH} "$CONDA_FILE"
-
-    else
-        gpuci_logger "Building cuml python in gpu job"
-        "$WORKSPACE/build.sh" -v cuml --codecov
-    fi
+    gpuci_logger "Building and installing cuml"
+    export CONDA_BLD_DIR="$WORKSPACE/.conda-bld"
+    export VERSION_SUFFIX=""
+    gpuci_conda_retry build --no-build-id --croot ${CONDA_BLD_DIR} conda/recipes/cuml -c ${CONDA_ARTIFACT_PATH} --python=${PYTHON}
+    gpuci_mamba_retry install -c ${CONDA_ARTIFACT_PATH} -c ${CONDA_BLD_DIR} cuml
 
     gpuci_logger "Install the main version of dask, distributed, and dask-glm"
     set -x
@@ -206,17 +199,11 @@ else
     set +x
 
     gpuci_logger "Python pytest for cuml"
-    cd $WORKSPACE/python
+    cd $WORKSPACE/python/cuml/tests
 
-    # When installing cuml with project flash, we need to delete all folders except
-    # cuml/tests since we are not building cython extensions in place
-    if [ "$PYTHON" == "3.8" ];then
-        find ./cuml -mindepth 1 ! -regex '^./cuml/tests\(/.*\)?' -delete
-    fi
+    pytest --cache-clear --basetemp=${WORKSPACE}/cuml-cuda-tmp --junitxml=${WORKSPACE}/junit-cuml.xml -v -s -m "not memleak" --durations=50 --timeout=300 --ignore=dask --cov-config=.coveragerc --cov=cuml --cov-report=xml:${WORKSPACE}/python/cuml/cuml-coverage.xml --cov-report term
 
-    pytest --cache-clear --basetemp=${WORKSPACE}/cuml-cuda-tmp --junitxml=${WORKSPACE}/junit-cuml.xml -v -s -m "not memleak" --durations=50 --timeout=300 --ignore=cuml/tests/dask --ignore=cuml/raft --cov-config=.coveragerc --cov=cuml --cov-report=xml:${WORKSPACE}/python/cuml/cuml-coverage.xml --cov-report term
-
-    timeout 7200 sh -c "pytest cuml/tests/dask --cache-clear --basetemp=${WORKSPACE}/cuml-mg-cuda-tmp --junitxml=${WORKSPACE}/junit-cuml-mg.xml -v -s -m 'not memleak' --durations=50 --timeout=300 --cov-config=.coveragerc --cov=cuml --cov-report=xml:${WORKSPACE}/python/cuml/cuml-dask-coverage.xml --cov-report term"
+    timeout 7200 sh -c "pytest dask --cache-clear --basetemp=${WORKSPACE}/cuml-mg-cuda-tmp --junitxml=${WORKSPACE}/junit-cuml-mg.xml -v -s -m 'not memleak' --durations=50 --timeout=300 --cov-config=.coveragerc --cov=cuml --cov-report=xml:${WORKSPACE}/python/cuml/cuml-dask-coverage.xml --cov-report term"
 
     ################################################################################
     # TEST - Run notebook tests
